breaking up the files before I get too deep into adding the forge installer
